+++ Package.swift
@@ -15,7 +15,7 @@ let package = Package(
         .package(
             name: "SendbirdChatSDK",
             url: "https://github.com/sendbird/sendbird-chat-sdk-ios",
-            from: "4.19.9"
+            from: "4.20.0"
         ),
     ],
     targets: [