+++ Sources/View/Common/Menu/SBUMenuView.swift
@@ -264,7 +264,7 @@ class SBUMenuView: NSObject {
         var titleLabelPosX: CGFloat = 0
         var titleLabelWidth: CGFloat = 0
         var textAlignment: NSTextAlignment = .left
-        if UIView.appearance().semanticContentAttribute == .forceLeftToRight {
+        if UIView.getCurrentLayoutDirection().isLTR == true {
             textAlignment = .left
             titleLabelPosX = leftMargin
             if item.image != nil {